Refactor config handling and cleanup
