+++ Dockerfile
@@ -9,12 +9,12 @@ ENV GOARCH amd64
 
 RUN go mod tidy && go test ./... && go build -trimpath -ldflags="-w -s" -o rss_exporter
 
-FROM alpine:latest AS runner
+FROM registry-maas.maas.services.sabio.co.uk/docker/busybox-glibc:1.0.0
 
-WORKDIR /rss_exporter
+WORKDIR /
 COPY --from=builder /rss_exporter/rss_exporter .
-COPY --from=builder /rss_exporter/config.example.yml config.yml
 
 EXPOSE 9091/tcp
 
-ENTRYPOINT ["./rss_exporter"]
+ENTRYPOINT ["/rss_exporter"]
+CMD ["-config.file=/config/config.yml"]
