--- collectors/init.go
@@ -1,22 +1,0 @@
-package collectors
-
-import (
-	"os"
-	"time"
-
-	"github.com/sirupsen/logrus"
-)
-
-func init() {
-	logrus.SetLevel(logrus.InfoLevel)
-	logrus.SetOutput(os.Stdout)
-	logrus.SetFormatter(&logrus.TextFormatter{
-		TimestampFormat:        time.RFC3339,
-		DisableColors:          true,
-		DisableLevelTruncation: true,
-		ForceQuote:             true,
-		FullTimestamp:          true,
-	})
-
-	initConfig()
-}
