+++ docs/metrics.md
@@ -6,7 +6,6 @@ The exporter exposes the following Prometheus metrics:
 |--------|--------|-------------|
 | `rss_exporter_service_status` | `service`, `customer` (optional), `state` | Current service state: `ok`, `service_issue`, or `outage`. |
 | `rss_exporter_service_issue_info` | `service`, `customer` (optional), `service_name` (optional), `region` (optional), `title`, `link`, `guid` | Information about the active incident, value is always `1` when present. |
-| `rss_exporter_fetch_errors_total` | `service`, `customer` (optional) | Counter of consecutive feed fetch failures. |
 
 The `service_name` and `region` labels are only populated for providers that
 include this information in their feeds, such as **aws** and **azure**.
