+++ internal/fetcher/fetcher.go
@@ -1,4 +1,4 @@
-package connector
+package fetcher
 
 import (
 	"context"
